Was hanging regardless of answer to Proceed?, with connection left open
Thanks for the project! Trying to use it for our 250GB SQL Server DB, no luck so far. Took me a while to figure out the dependencies to use and the connection string, so that would really promote adoption for casual users.

Not 100% sure if this is the desired behavior, but it was hanging for me. Ctrl-C and Ctrl-Z did not halt it. There is probably something else wrong with my databases, users, or command - still haven't got it working - so see if it works as expected for you.

Command:
rdbms-subsetter --loglevel INFO --schema nvpers_dev_sm mssql+pymssql://sa:password1@localhost:1433/nvpers_dev mssql+pymssql://sa:password1@localhost:1433/nvpers_dev_sm 0.05

Stack: 
Windows7
Python 3.4.2
pip (1.5.6)
pymssql (2.1.1)
rdbms-subsetter (0.2.1)
setuptools (2.1)
SQLAlchemy (0.9.8)